+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -13,7 +13,7 @@ Version 3.1.4
 
  * 2014-04-06 Fixed bug (#165) that caused that html template caching did 
               not work properly. (Nomi + ra)
- * 2014-03-28 Added support for static Result.NO_HTTP_CONTENT. No need
+ * 2014-03-28 Added support for static Result.NO_HTTP_BODY. No need
               to create new() class all the time. (ra)
 
 Version 3.1.3